bugfix: sulist packages fail to be added due to missing reboot
added a reboot_redroid function after exeggcute_sulist function is called. sulist packages added via cli will fail to be enforced unless redroid is rebooted. this fixes the bug where exeggcute runs getenforce and receives an empty response. this is due to exeggcute being granted root permissions but not enabled via sulist enforcement. 

additional functions were added to verify this stage of the process:  check_magisk_sulist to verify the status of the functions were enabled properly, and check_exeggcute_sulist to verify the packages exist in the magisk db.